+++ changelog.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - Fix auto detection of libraries
+- Fix react settings
 
 ## [0.1.1] - 2023-03-07